# public/swagger/swagger-doc.json
{"openapi":"2.0.0","info":{"description":"数据管理平台","version":"2.0.0","title":"数据管理平台-接口文档","contact":{"name":"xupeng","email":""},"termsOfService":"http://www.yczcjk.com/","license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"host":"localhost:3000","basePath":"/api","paths":{"/auth/getCaptcha":{"get":{"tags":["登录"],"summary":"获取图形验证码","operationId":"getCaptcha","parameters":[{"in":"query","name":"phone","required":true,"type":"string","description":"手机号"}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/CaptchaResponse"}}}}},"/auth/preLogin":{"get":{"tags":["登录"],"summary":"检查是否需要验证码","operationId":"preLogin","parameters":[{"in":"query","name":"phone","required":true,"type":"string","description":"手机号"}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/PreLoginResponse"}}}}},"/auth/login":{"post":{"tags":["登录"],"summary":"登录","operationId":"login","produces":["application/json"],"parameters":[{"in":"body","name":"request","required":true,"schema":{"$ref":"#/definitions/LoginRequest"}}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/LoginResponse"}}}}},"/auth/logout":{"get":{"tags":["登录"],"summary":"退出登录","operationId":"logout","produces":["application/json"],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/EmptyResponse"}}}}},"/auth/getUserInfo":{"get":{"tags":["登录"],"summary":"获取用户信息","operationId":"getUserInfo","produces":["application/json"],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/UserInfo"}}}}},"/auth/editPassword":{"post":{"tags":["登录"],"summary":"修改密码","operationId":"editPassword","produces":["application/json"],"parameters":[{"in":"body","name":"request","required":true,"schema":{"$ref":"#/definitions/EditPasswordRequest"}}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/EmptyResponse"}}}}},"/admin/addAccount":{"post":{"tags":["账号管理"],"summary":"添加账号","operationId":"addAccount","produces":["application/json"],"parameters":[{"in":"body","name":"request","required":true,"schema":{"$ref":"#/definitions/AddAccountRequest"}}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/EmptyResponse"}},"5005":{"description":"账号已存在"}}}},"/admin/deleteAccount":{"post":{"tags":["账号管理"],"summary":"删除账号","operationId":"deleteAccount","produces":["application/json"],"parameters":[{"in":"body","name":"request","required":true,"schema":{"$ref":"#/definitions/DeleteAccountRequest"}}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/EmptyResponse"}}}}},"/admin/resetPassword":{"post":{"tags":["账号管理"],"summary":"重置密码","operationId":"resetPassword","produces":["application/json"],"parameters":[{"in":"body","name":"request","required":true,"schema":{"$ref":"#/definitions/ResetPasswordRequest"}}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/EmptyResponse"}}}}},"/admin/getUserList":{"get":{"tags":["账号管理"],"summary":"获取用户列表","produces":["application/json"],"operationId":"getUserList","parameters":[{"in":"query","name":"role","required":true,"type":"integer","format":"int32","description":"角色（1-开发账号  2-测试账号）"},{"in":"query","name":"num","required":true,"type":"integer","format":"int32","description":"每页条数，默认10 ,示例值(10)"},{"in":"query","name":"page","required":true,"type":"integer","format":"int32","description":"页数。1开始，默认1 ,示例值(1)"}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/UserListResponse"}}}}},"/abnormal/getHandler":{"get":{"tags":["账号管理"],"summary":"查询账号人员数据","operationId":"getHandler","produces":["application/json"],"parameters":[{"in":"query","name":"role","required":true,"type":"integer","format":"int32","description":"权限"},{"in":"query","name":"isDeleted","required":false,"type":"integer","format":"int32","description":"是否包含已删除账号（0-不包含 1-包含）"}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/UserResponse"}}}}},"/abnormal/getDataType":{"get":{"tags":["开发异常"],"summary":"获取数据类型","operationId":"getDataType","produces":["application/json"],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/AbnormalTypeResponse"}}}}},"/abnormal/getAbnormalList":{"post":{"tags":["开发异常"],"summary":"查询开发异常列表","operationId":"getAbnormalList","produces":["application/json"],"parameters":[{"in":"body","name":"request","required":true,"schema":{"$ref":"#/definitions/AbnormalListRequest"}}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/AbnormalListResponse"}}}}},"/abnormal/getAbnormalInfo":{"get":{"tags":["开发异常"],"summary":"查询开发异常详情","operationId":"getAbnormalInfo","produces":["application/json"],"parameters":[{"in":"query","name":"id","required":true,"type":"integer","format":"int32","description":"主键id"}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/AbnormalInfoResponse"}}}}},"/abnormal/handleAbnormal":{"post":{"tags":["开发异常"],"summary":"处理开发异常","operationId":"handleAbnormal","produces":["application/json"],"parameters":[{"in":"body","name":"request","required":true,"schema":{"$ref":"#/definitions/HandleAbnormalRequest"}}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/EmptyResponse"}}}}},"/mapping/getTableName":{"get":{"tags":["新增映射值"],"summary":"查询表名","operationId":"getTableName","produces":["application/json"],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/TableNameResponse"}}}}},"/mapping/getFieldName":{"get":{"tags":["新增映射值"],"summary":"查询表字段名","operationId":"getFieldName","produces":["application/json"],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/FieldNameResponse"}}}}},"/mapping/getRelation":{"get":{"tags":["新增映射值"],"summary":"查询映射关系","operationId":"getRelation","produces":["application/json"],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/RelationResponse"}}}}},"/mapping/getMappingList":{"post":{"tags":["新增映射值"],"summary":"查询列表数据","operationId":"getMappingList","produces":["application/json"],"parameters":[{"in":"body","name":"request","required":true,"schema":{"$ref":"#/definitions/MappingListRequest"}}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/MappingListResponse"}}}}},"/mapping/addMapping":{"post":{"tags":["新增映射值"],"summary":"添加映射关系","operationId":"addMapping","produces":["application/json"],"parameters":[{"in":"body","name":"request","required":true,"schema":{"$ref":"#/definitions/AddMappingRequest"}}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/EmptyResponse"}}}}},"/mapping/getHistoryMapping":{"get":{"tags":["新增映射值"],"summary":"查看历史映射","operationId":"getHistoryMapping","produces":["application/json"],"parameters":[{"in":"body","name":"id","required":true,"type":"integer","format":"int32","description":"数据id"}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/HistoryMappingResponse"}}}}},"/mapping/upload":{"get":{"tags":["新增映射值"],"summary":"上传","operationId":"upload","produces":["application/json"],"parameters":[{"in":"body","name":"idList","type":"string","required":true,"description":"数据id（示例:\"1,2,3\"）"}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/EmptyResponse"}}}}},"/mapping/back":{"get":{"tags":["新增映射值"],"summary":"退回","operationId":"back","produces":["application/json"],"parameters":[{"in":"body","name":"id","type":"integer","format":"int32","required":true,"description":"数据id"}],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/EmptyResponse"}}}}},"/task/updateAbnormal":{"get":{"tags":["定时任务"],"summary":"更新开发异常数据","operationId":"abnormal","produces":["application/json"],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/EmptyResponse"}}}}},"/task/updateMapping":{"get":{"tags":["定时任务"],"summary":"更新映射值数据","operationId":"mapping","produces":["application/json"],"responses":{"200":{"description":"success","schema":{"$ref":"#/definitions/EmptyResponse"}}}}}},"definitions":{"EmptyResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32","description":"返回码,200为正常,详情参考http状态码"},"data":{"type":"object","description":"返回数据"},"message":{"type":"string","description":"返回描述信息"}},"xml":{"name":"EmptyResponse"}},"UserResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32","default":200,"description":"返回码,200为正常,详情参考http状态码"},"data":{"$ref":"#/definitions/UserData"},"message":{"type":"string","description":"返回描述信息"}},"xml":{"name":"UserResponse"}},"UserData":{"type":"object","properties":{"list":{"type":"array","xml":{"name":"list","wrapped":true},"items":{"$ref":"#/definitions/UserDataItem"}}},"xml":{"name":"UserData"}},"UserDataItem":{"type":"object","properties":{"id":{"type":"inetger","format":"int32","description":"id"},"name":{"type":"string","description":"姓名"}},"xml":{"name":"UserDataItem"}},"CaptchaResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32","default":200,"description":"返回码,200为正常,详情参考http状态码"},"data":{"$ref":"#/definitions/CaptchaData"},"message":{"type":"string","description":"返回描述信息"}},"xml":{"name":"CaptchaResponse"}},"CaptchaData":{"type":"object","properties":{"captcha":{"type":"string","description":"图片验证码（base64）"}},"xml":{"name":"CaptchaData"}},"PreLoginResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32","default":200,"description":"返回码,200为正常,详情参考http状态码"},"data":{"$ref":"#/definitions/PreLoginData"},"message":{"type":"string","description":"返回描述信息"}},"xml":{"name":"PreLoginResponse"}},"PreLoginData":{"type":"object","properties":{"needCode":{"type":"boolean","description":"是否需要图片验证码（true-是 false-否）"}},"xml":{"name":"PreLoginData"}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string","description":"手机号"},"password":{"type":"string","description":"密码"},"code":{"type":"string","description":"验证码"}},"xml":{"name":"LoginRequest"}},"LoginResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32","default":200,"description":"返回码,200为正常,详情参考http状态码"},"data":{"$ref":"#/definitions/UserInfo"},"message":{"type":"string","description":"返回描述信息"}},"xml":{"name":"LoginResponse"}},"UserInfo":{"type":"object","properties":{"id":{"type":"integer","description":"用户id"},"role":{"type":"integer","description":"角色（0-管理账号  1-开发账号  2-测试账号）"},"user_name":{"type":"string","description":"姓名"},"user_mobile":{"type":"string","description":"手机号"},"is_modified":{"type":"integer","description":"是否需要修改密码（0-是 1-否）"}},"xml":{"name":"LoginData"}},"EditPasswordRequest":{"type":"object","properties":{"newPassword":{"type":"string","description":"新密码"},"oldPassword":{"type":"string","description":"旧密码"}},"xml":{"name":"EditPasswordRequest"}},"ResetPasswordRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int32","description":"用户id"}},"xml":{"name":"ResetPasswordRequest"}},"AddAccountRequest":{"type":"object","properties":{"role":{"type":"integer","format":"int32","description":"角色（1-开发账号  2-测试账号）"},"name":{"type":"string","description":"姓名"},"password":{"type":"string","description":"密码"},"phone":{"type":"string","description":"手机号"}},"xml":{"name":"AddAccountRequest"}},"DeleteAccountRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int32","description":"用户id"}},"xml":{"name":"DeleteAccountRequest"}},"UserListResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32","default":200,"description":"返回码,200为正常,详情参考http状态码"},"data":{"$ref":"#/definitions/UserListData"},"message":{"type":"string","description":"返回描述信息"}},"xml":{"name":"UserListResponse"}},"UserListData":{"type":"object","properties":{"id":{"type":"integer","format":"int32","description":"用户id"},"name":{"type":"string","description":"姓名"},"phone":{"type":"string","description":"手机号"},"abnormalNum":{"type":"integer","format":"int64","description":"处理中开发异常数"},"editField":{"type":"integer","format":"int64","description":"待修改映射数"}},"xml":{"name":"UserListData"}},"AbnormalTypeResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32","default":200,"description":"返回码,200为正常,详情参考http状态码"},"data":{"$ref":"#/definitions/AbnormalType"},"message":{"type":"string","description":"返回描述信息"}},"xml":{"name":"UserListResponse"}},"AbnormalType":{"type":"object","properties":{"dataType":{"type":"array","xml":{"name":"dataType","wrapped":true},"items":{"$ref":"#/definitions/AbnormalTypeItem"}}},"xml":{"name":"AbnormalType"}},"AbnormalTypeItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64","description":"数据类型id"},"typeName":{"type":"string","description":"数据类型名称"}},"xml":{"name":"AbnormalTypeItem"}},"AbnormalListRequest":{"type":"object","properties":{"num":{"type":"integer","format":"int32","description":"每页条数，默认10 ,示例值(10)"},"page":{"type":"integer","format":"int32","description":"页数。1开始，默认1 ,示例值(1)"},"sortColumn":{"type":"string","description":"排序字段（PUBLISHTIME、FAILEDTIME、PRIORITYTIME、HANDLETIME）"},"sortOrder":{"type":"string","description":"排序顺序（ASC,DESC）"},"handleStatus":{"type":"integer","format":"int32","description":"处理状态(0-待处理 1-处理中 2-已处理)"},"dataType":{"type":"integer","format":"int64","description":"数据类型id"},"sourceId":{"type":"integer","format":"int64","description":"source_id"},"sourceName":{"type":"string","description":"网站名称"},"priority":{"type":"integer","format":"int64","description":"优先级（1-13）"},"failedType":{"type":"integer","format":"int32","description":"错误类型（1-解析错误 2-抓取错误）"},"handler":{"type":"integer","format":"int64","description":"处理人id"},"publishTimeStart":{"type":"string","format":"date","description":"发布日期（开始日期），示例值(2021-01-01)"},"publishTimeEnd":{"type":"string","format":"date","description":"发布日期（结束日期），示例值(2021-01-01)"},"failedTimeStart":{"type":"string","format":"date","description":"报错日期（开始日期），示例值(2021-01-01)"},"failedTimeEnd":{"type":"string","format":"date","description":"报错日期（结束日期），示例值(2021-01-01)"},"handleTimeStart":{"type":"string","format":"date","description":"处理日期（开始日期），示例值(2021-01-01)"},"handleTimeEnd":{"type":"string","format":"date","description":"处理日期（结束日期），示例值(2021-01-01)"}},"xml":{"name":"AbnormalListRequest"}},"AbnormalListResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32","description":"id"},"gmtPublish":{"type":"string","format":"date","description":"发布日期，示例值(2021-01-01)"},"gmtFailed":{"type":"string","format":"date","description":"报错日期，示例值(2021-01-01)"},"gmtHandle":{"type":"string","format":"date","description":"处理日期，示例值(2021-01-01)"},"priority":{"type":"integer","format":"int64","description":"优先级"},"sourceType":{"type":"string","description":"数据类型"},"failedType":{"type":"string","description":"错误类型"},"sourceId":{"type":"integer","format":"int64","description":"source_id"},"sourceName":{"type":"string","description":"网站名称"},"sourceUrl":{"type":"string","description":"网站链接"},"childCatalog":{"type":"string","description":"子目录"},"catalogUrl":{"type":"string","description":"子目录链接"},"crawlerName":{"type":"string","description":"爬虫名称"},"handler":{"type":"string","description":"处理人"},"isHandlerDeleted":{"type":"integer","format":"int32","description":"处理人是否已删除(0-未删除 1-已删除)"}},"xml":{"name":"AbnormalListResponse"}},"AbnormalInfoResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32","default":200,"description":"返回码,200为正常,详情参考http状态码"},"data":{"$ref":"#/definitions/AbnormalInfoData"},"message":{"type":"string","description":"返回描述信息"}},"xml":{"name":"AbnormalInfoResponse"}},"AbnormalInfoData":{"type":"object","properties":{"sourceId":{"type":"integer","format":"int64","description":"source_id"},"sourceName":{"type":"string","description":"网站名称"},"childCatalog":{"type":"string","description":"子目录"},"crawlerName":{"type":"string","description":"爬虫名称"},"failedContent":{"type":"string","description":"报错内容"},"remarks":{"type":"string","description":"备注"},"failedHistory":{"type":"array","description":"报错历史","xml":{"name":"time","wrapped":true},"items":{"type":"string","format":"data"}},"handleLog":{"type":"array","description":"处理记录","xml":{"name":"log","wrapper":true},"items":{"$ref":"#/definitions/HandleLog"}}},"xml":{"name":"AbnormalInfoData"}},"HandleLog":{"type":"object","properties":{"handleTime":{"type":"string","format":"data","description":"处理时间"},"handler":{"type":"string","description":"处理人"},"remarks":{"type":"string","description":"备注内容"}},"xml":{"name":"HandleLog"}},"HandleAbnormalRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int32","description":"id"},"handleStatus":{"type":"integer","format":"int32","description":"处理状态（1-处理中 2-已处理）"},"remarks":{"type":"string","description":"备注"}},"xml":{"name":"HandleAbnormalRequest"}},"TableNameResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32","default":200,"description":"返回码,200为正常,详情参考http状态码"},"data":{"$ref":"#/definitions/TableNameData"},"message":{"type":"string","description":"返回描述信息"}},"xml":{"name":"TableNameResponse"}},"TableNameData":{"type":"object","properties":{"tableList":{"type":"array","description":"表名列表","xml":{"name":"tableName","wrapper":true},"items":{"$ref":"#/definitions/TableNameItem"}}}},"TableNameItem":{"type":"object","properties":{"id":{"type":"integer","format":"int32","description":"id"},"tableName":{"type":"string","description":"表名"}},"xml":{"name":"TableNameItem"}},"FieldNameResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32","default":200,"description":"返回码,200为正常,详情参考http状态码"},"data":{"$ref":"#/definitions/FieldNameData"},"message":{"type":"string","description":"返回描述信息"}},"xml":{"name":"FieldNameResponse"}},"FieldNameData":{"type":"object","properties":{"fieldList":{"type":"array","description":"表名列表","xml":{"name":"FieldNameData","wrapper":true},"items":{"$ref":"#/definitions/FieldNameItem"}}}},"FieldNameItem":{"type":"object","properties":{"id":{"type":"integer","format":"int32","description":"id"},"fieldName":{"type":"string","description":"表字段名"}},"xml":{"name":"FieldNameItem"}},"RelationResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32","default":200,"description":"返回码,200为正常,详情参考http状态码"},"data":{"$ref":"#/definitions/RelationData"},"message":{"type":"string","description":"返回描述信息"}},"xml":{"name":"RelationResponse"}},"RelationData":{"type":"object","properties":{"relationList":{"type":"array","description":"表名列表","xml":{"name":"RelationData","wrapper":true},"items":{"$ref":"#/definitions/RelationItem"}}}},"RelationItem":{"type":"object","properties":{"result":{"type":"string","description":"映射结果"},"code":{"type":"integer","format":"int64","description":"映射码值"}},"xml":{"name":"RelationItem"}},"MappingListRequest":{"type":"object","properties":{"status":{"type":"integer","format":"int32","description":"处理状态（1-待处理  2-已处理、待测试  3-已退回、待修改  4-已修改  5-已上传）"},"tableName":{"type":"string","description":"表名"},"fieldName":{"type":"string","description":"字段名"},"result":{"type":"string","description":"映射结果"},"code":{"type":"integer","format":"int64","description":"映射码值"},"handlerId":{"type":"integer","format":"int32","description":"处理人id"},"testId":{"type":"integer","format":"int32","description":"测试人id"},"backTimeStart":{"type":"string","format":"data","description":"退回时间-开始时间"},"backTimeEnd":{"type":"string","format":"data","description":"退回时间-结束时间"},"handleTimeStart":{"type":"string","format":"data","description":"处理时间-开始时间"},"handleTimeEnd":{"type":"string","format":"data","description":"处理时间-结束时间"},"createTimeStart":{"type":"string","format":"data","description":"发布时间-开始时间"},"createTimeEnd":{"type":"string","format":"data","description":"发布时间-结束时间"},"updateTimeStart":{"type":"string","format":"data","description":"更新时间-开始时间"},"updateTimeEnd":{"type":"string","format":"data","description":"更新时间-结束时间"},"sortColumn":{"type":"string","description":"排序字段（UPDATETIME、HANDLETIME、BACKTIME、UPLOADTIME）"},"sortOrder":{"type":"string","description":"排序顺序（ASC,DESC）"}}},"MappingListResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32","default":200,"description":"返回码,200为正常,详情参考http状态码"},"data":{"$ref":"#/definitions/MappingListData"},"message":{"type":"string","description":"返回描述信息"}},"xml":{"name":"MappingListResponse"}},"MappingListData":{"type":"object","properties":{"createTime":{"type":"string","format":"data","description":"首次发布时间"},"modifiedTime":{"type":"string","format":"data","description":"更新时间"},"handleTime":{"type":"string","format":"data","description":"处理时间"},"backTime":{"type":"string","format":"data","description":"退回时间"},"uploadTime":{"type":"string","format":"data","description":"上传时间"},"tableName":{"type":"string","description":"表名"},"mapField":{"type":"string","description":"表字段名"},"fieldName":{"type":"string","description":"新增名称"},"resule":{"type":"string","description":"映射结果"},"code":{"type":"integer","format":"int64","description":"映射码值"},"handler":{"type":"string","description":"处理人"},"isHandlerDeleted":{"type":"string","description":"处理人是否删除（0-未删除 1-已删除）"},"tester":{"type":"integer","format":"int 32","description":"测试人员"},"isTesterDeleted":{"type":"integer","format":"int 32","description":"测试人员是否删除（0-未删除 1-已删除）"}}},"AddMappingRequest":{"type":"object","properties":{"result":{"type":"string","description":"映射名称"},"code":{"type":"integer","format":"int64","description":"映射码值"},"tableName":{"type":"string","description":"表名"},"mapField":{"type":"string","description":"表字段名"},"fieldName":{"type":"string","description":"新增名称"}},"xml":{"name":"AddMappingRequest"}},"HistoryMappingResponse":{"type":"object","properties":{"tableName":{"type":"string","description":"表名"},"mapField":{"type":"string","description":"字段名"},"list":{"type":"array","xml":{"name":"list","wrapped":true},"items":{"$ref":"#/definitions/HistoryMappingItem"}}}},"HistoryMappingItem":{"type":"object","properties":{"primitive":{"type":"string","description":"原始值"},"result":{"type":"string","description":"映射结果"},"code":{"type":"integer","format":"int64","description":"映射码值"}},"xml":{"name":"HistoryMappingItem"}}},"components":{},"tags":[]}